backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.7.1
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
Werkzeug==2.3.7
numpy==1.24.3
scipy==1.11.4
astropy==5.3.4
skyfield==1.46
ephem==4.1.4
matplotlib==3.7.2
Pillow==10.0.1
pandas==2.0.3
geopy==2.4.0
beautifulsoup4==4.12.2
lxml==4.9.3
schedule==1.2.0
websockets==11.0.3
aiohttp==3.8.6
asyncio==3.4.3
eventlet==0.33.3
gunicorn==21.2.0